Add various test 'puts' commands to solve bug of..
board_array not updating correctly and a bug in #promote_pawn
due to not identifying the class of the output of the class
command in the case command of #update_the_board. Also to
insert 'self' several times in Board class to make sure that
instance variables are being changed.