Add @colour to ChangeTheBoard. Add ..
opponent_back_rank clause to #make_general_move and add optional
argument to #remove_castling_rights in Board class so that when
rooks are captured in their starting positions, castling rights
are lost on that side for the opponent. Temporarily change NEW_BOARD_ARRAY
to check number of legal moves found.