Add #any_en_passent_in_theory? to Board class
so that Game in #one_turn can precisely determine if there
are any legal en_passent moves and, if not, tell the Board
to reset its @en_passent variable, which was updated when a
pawn moved 2 squares forwards, JUST IN CASE an enemy pawn could
capture it en_passent next move.
However, for 3 fold repetition to be triggered correctly, the Board
object should ONLY be stored with en_passent possibilities when there
is such a capture legally possible on the next move.