Add Goodbye message loop to top level code.
Solve bug in which the constant array NEW_BOARD_ARRAY was still
somehow finding shadows of pieces that had moved in a previous game
(although only pawns were affected!). This meant that when playing e2e4 or
some other pawn move and saving the game, if we started a new game, the same
pawn would be missing one or both legal moves. The change is:
Replace NEW_BOARD_ARRAY with an explicit non-constant array in Board class.