Add computer player option, which captures ..
high-valued pieces where possible, or makes other captures, or just
chooses a move at random. It won't know about checks or checkmate though.
Refactored ChangeTheBoard class to be instantiated with instance variable @player so
that it can ask the player object for which piece to promote a pawn to. That will allow
computer and human players to respond differently. Also remove testing 'puts' statements
throughout the files of /lib.