Refactor #toggle_colours in Game class to toggle..
the colours in Game and Board classes together to avoid confusion. Refactor
arguments of various methods in #consequences_of_move to allow for @moving_name
and @not_moving_name being interchanged.
Add feature of declaring Check and reminding us who the players are
and if the side to move is in Check when reloading a saved game using
 #name_the_players, #remind_check and #declare_check.
Minor style changes in GenerateLegalMoves.