Add GameInputs class to allow inputs..
for a Game to be entered at the start for creating the players and entering moves, to help with testing. Default is for no inputs supplied, so gets to be used then as before. Add #input_to_use methods to Game and Player classes. Will not apply to Computer players or saving/loading games.